Split file and check for return
